+++ src/Device/DeviceSelector/DeviceSelector.tsx
@@ -8,18 +8,21 @@ import React, { useCallback, useEffect, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { DeviceTraits } from '@nordicsemiconductor/nrf-device-lib-js';
 
-import { Device } from '../../state';
 import useHotKey from '../../utils/useHotKey';
 import {
-    clearWaitForDevice,
     getWaitingToAutoReselect,
     setAutoSelectDevice,
 } from '../deviceAutoSelectSlice';
-import { startWatchingDevices, stopWatchingDevices } from '../deviceLister';
+import {
+    clearWaitForDevice,
+    startWatchingDevices,
+    stopWatchingDevices,
+} from '../deviceLister';
 import { DeviceSetupConfig, setupDevice } from '../deviceSetup';
 import DeviceSetupView from '../DeviceSetup/DeviceSetupView';
 import {
     deselectDevice,
+    Device,
     deviceIsSelected as deviceIsSelectedSelector,
     selectDevice,
     selectedSerialNumber,