+++ src/Device/deviceAutoSelectSlice.ts
@@ -6,12 +6,32 @@
 
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 
-import {
-    Device,
-    DeviceAutoSelectState,
-    RootState,
-    WaitForDevice,
-} from '../state';
+import type { RootState } from '../store';
+import type { Device } from './deviceSlice';
+
+export interface WaitForDevice {
+    timeout: number;
+    when:
+        | 'always'
+        | 'applicationMode'
+        | 'dfuBootLoaderMode'
+        | 'sameTraits'
+        | ((device: Device) => boolean);
+    once: boolean;
+    onSuccess?: (device: Device) => void;
+    onFail?: (reason?: string) => void;
+}
+
+export interface DeviceAutoSelectState {
+    autoReselect: boolean;
+    device?: Device;
+    disconnectionTime?: number;
+    waitForDevice?: WaitForDevice;
+    autoReconnectTimeout?: NodeJS.Timeout;
+    lastArrivedDeviceId?: number;
+    arrivedButWrongWhen?: boolean;
+    onCancelTimeout?: () => void;
+}
 
 const initialState: DeviceAutoSelectState = {
     autoReselect: false,
@@ -29,9 +49,17 @@ const slice = createSlice({
             state.autoReconnectTimeout = action.payload;
         },
 
-        clearWaitForDeviceTimeout: state => {
+        clearWaitForDeviceTimeout: (
+            state,
+            { payload: clearDevice }: PayloadAction<boolean>
+        ) => {
             clearTimeout(state.autoReconnectTimeout);
             state.autoReconnectTimeout = undefined;
+
+            if (clearDevice) {
+                state.onCancelTimeout = undefined;
+                state.waitForDevice = undefined;
+            }
         },
 
         setAutoSelectDevice: (
@@ -64,17 +92,20 @@ const slice = createSlice({
             if (state.device) state.waitForDevice = action.payload;
         },
 
-        clearWaitForDevice: state => {
-            state.waitForDevice = undefined;
-            clearTimeout(state.autoReconnectTimeout);
-            state.autoReconnectTimeout = undefined;
-        },
-
         setLastArrivedDeviceId: (
             state,
             action: PayloadAction<number | undefined>
         ) => {
             if (state.device) state.lastArrivedDeviceId = action.payload;
+        },
+        setArrivedButWrongWhen: (
+            state,
+            action: PayloadAction<boolean | undefined>
+        ) => {
+            state.arrivedButWrongWhen = action.payload;
+        },
+        setOnCancelTimeout: (state, action: PayloadAction<() => void>) => {
+            state.onCancelTimeout = action.payload;
         },
     },
 });
@@ -88,8 +119,9 @@ export const {
         setDisconnectedTime,
         setAutoReselect,
         setWaitForDevice,
-        clearWaitForDevice,
         setLastArrivedDeviceId,
+        setArrivedButWrongWhen,
+        setOnCancelTimeout,
     },
 } = slice;
 
@@ -114,3 +146,5 @@ export const getWaitForDevice = (state: RootState) =>
 
 export const getLastArrivedDeviceId = (state: RootState) =>
     state.deviceAutoSelect.lastArrivedDeviceId;
+export const getArrivedButWrongWhen = (state: RootState) =>
+    state.deviceAutoSelect.arrivedButWrongWhen;